requirements.txt
alembic==1.12.1
bcrypt==4.0.1
blinker==1.6.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysql-connector-python-rf==2.2.2
packaging==23.2
protobuf==4.21.12
psycopg2==2.9.9
pycparser==2.21
PyMySQL==1.1.0
setuptools==68.2.2
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.0
WTForms==3.1.0
